+++ enzyme/tools/enzyme-tblgen/blas-tblgen.h
@@ -1,0 +1,2 @@
+void emitBlasDerivatives(const llvm::RecordKeeper &RK, llvm::raw_ostream &os);
+bool hasDiffeRet(llvm::Init *resultTree);
